requirements.txt
# Web Framework
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
pydantic-settings==2.0.3
python-multipart==0.0.6
email-validator==2.0.0
slowapi==0.1.9

# Database
sqlalchemy==2.0.20
alembic==1.12.0
psycopg2-binary==2.9.7
asyncpg==0.28.0
greenlet==2.0.2

# Authentication (Supabase client)
supabase==1.0.3
python-jose==3.3.0
cryptography==41.0.3

# Background Tasks
celery==5.3.1
redis==4.6.0
flower==2.0.1
apscheduler==3.10.4

# API Integrations
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0

fitbit==0.3.1
oauthlib==3.2.2
requests-oauthlib==1.3.1

openai==0.28.1
plaid-python==14.1.0

# Data Processing
pandas==2.1.0
numpy==1.25.2
scikit-learn==1.3.0
python-dateutil==2.8.2
statsmodels==0.14.0
matplotlib==3.7.3
plotly==5.16.1

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
httpx==0.23.3  # ✅ Downgraded for supabase compatibility
jinja2==3.1.2
pyyaml==6.0.1
ujson==5.8.0
aiofiles==23.2.1
bcrypt==4.0.1


# Testing
pytest==7.4.1
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.11.1
faker==19.6.1